+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Update OpenAI Embeddings model to support all allowed types of input [#6](https://github.com/hypermodeAI/models-as/pull/6)
 - More updates for OpenAI Embeddings model [#7](https://github.com/hypermodeAI/models-as/pull/7)
 - Add documentation comments for OpenAI Chat model [#8](https://github.com/hypermodeAI/models-as/pull/8)
+- Add experimental classification and embeddings models [#9](https://github.com/hypermodeAI/models-as/pull/9)
 
 ## 2024-06-28 - Version 0.1.6