+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-## UNRELEASED
+## 2024-06-24 - Version 0.2.1
 
 - Rename classes in OpenAI models [#11](https://github.com/hypermodeAI/models-as/pull/11)